fix: 자동 완성 기능
