+++ frontend/js/index.js
@@ -9,10 +9,7 @@ const headerEl = $(".header");
 const header = new HeaderView(headerEl);
 const searchEl = $(".search");
 const searchForm = new SearchFormView(searchEl);
-const recommendEl = $(".recommend");
-const autoComplete = new AutoCompleteView(recommendEl);
 const topBarEl = $(".top-bar");
 const topBar = new TopBarView(topBarEl);
 
-const c = new AutoCompleteController();
-console.log(c);
+const autoCompleteController = new AutoCompleteController();
